提取layout，，统一代码风格，优化路由
