+++ src/api/user.js
@@ -1,5 +1,4 @@
 import api from './index'
-
 // axios
 import request from '@/utils/request'
