+++ src/assets/css/index.scss
@@ -1,14 +1,13 @@
 @import './variables.scss';
 @import './mixin.scss';
- 
+
 html,
-body
-.app {
+body .app {
   color: #333333;
   font-family: Arial, Helvetica, 'STHeiti STXihei', 'Microsoft YaHei', Tohoma, sans-serif;
   background-color: $background-color;
 }
- 
-.app-container{
-  padding-bottom:50px;
+
+.app-container {
+  padding-bottom: 50px;
 }
