+++ src/main.js
@@ -8,15 +8,14 @@ import App from './App.vue'
 import router from './router'
 import store from './store'
 
-// 引入全局样式
-import '@/assets/css/index.scss'
 // 设置 js中可以访问 $cdn
 import { $cdn } from '@/config'
 Vue.prototype.$cdn = $cdn
 
 // 全局引入按需引入UI库 vant
 import '@/plugins/vant'
-
+// 引入全局样式
+import '@/assets/css/index.scss'
 // 移动端适配
 import 'lib-flexible/flexible.js'
