+++ vue.config.js
@@ -64,14 +64,13 @@ module.exports = {
     // }
   },
   css: {
-    extract: IS_PROD,
+    extract: IS_PROD, //是否将组件中的 CSS 提取至一个独立的 CSS 文件中 (而不是动态注入到 JavaScript 中的 inline 代码)。
     sourceMap: false,
     loaderOptions: {
       scss: {
         // 向全局sass样式传入共享的全局变量, $src可以配置图片cdn前缀
         // 详情: https://cli.vuejs.org/guide/css.html#passing-options-to-pre-processor-loaders
         prependData: `
-          @import "assets/css/index.scss";
           @import "assets/css/mixin.scss";
           @import "assets/css/variables.scss";
           $cdn: "${defaultSettings.$cdn}";
